+++ src/cpu/x64/jit_avx512_core_bf16_conv_kernel.hpp
@@ -159,8 +159,8 @@ private:
     }
 
     inline dim_t get_src_offset(dim_t ic_idx, dim_t isp) {
-        int icb = ic_idx / jcp.ic_block;
-        int ic = ic_idx % jcp.ic_block;
+        int icb = static_cast<int>(ic_idx / jcp.ic_block);
+        int ic = static_cast<int>(ic_idx % jcp.ic_block);
         dim_t isp_str = is_src_layout_nxc()
                 ? static_cast<dim_t>(jcp.ngroups) * jcp.ic
                 : (jcp.is_1stconv ? 1 : jcp.ic_block);
